+++ src/model/ConfigPreset.ts
@@ -88,10 +88,17 @@ export interface NodePreset extends DockerServicePreset {
     brokerExcludeDockerService?: boolean;
     brokerCompose?: any;
     brokerDockerComposeDebugMode?: boolean;
-    //super node
+
+    //Reward program
     rewardProgram?: string;
+    rewardProgramAgentIpv4_address?: string;
     rewardProgramAgentOpenPort?: boolean | number | string;
+    rewardProgramAgentExcludeDockerService?: boolean;
+    rewardProgramAgentCompose?: any;
+    rewardProgramAgentHost?: string;
+    rewardProgramAgentDockerComposeDebugMode?: boolean;
     agentUrl?: string; //calculated if not provided.
+
     restGatewayUrl?: string; // calculated if not provided;
 }
 
@@ -176,6 +183,7 @@ export interface ConfigPreset {
     symbolWalletImage: string;
     symbolFaucetImage: string;
     symbolServerImage: string;
+    symbolAgentBaseImage: string;
     symbolRestImage: string;
     votingKeyStartEpoch: number;
     votingKeyEndEpoch: number;